Pass the slice to the HoverBox
Collect current hovered items in the Canvas instead of the HoverBox.
When showing the HoverBox also inform it of the current slice. Allow a
way to get the allocated hours for a reservation from the slice. This
allows the HoverBox to have information about the allocated hours for a
slice.